+++ test/integration-tests/serverless-com-task.test.ts
@@ -14,6 +14,7 @@ describe('when calling org-formation perform tasks', () => {
     let spawnProcessAfterDeploy1Target: jest.MockContext<any, any>;
     let stateAfterDeploy1Target: GetObjectOutput;
     let spawnProcessAfterRerunFileWithoutChanges: jest.MockContext<any, any>;
+    let spawnProcessAfterRerunFileWithForceDeploy: jest.MockContext<any, any>;
     let spawnProcessAfterUpdateWithParams: jest.MockContext<any, any>;
     let stateAfterUpdateWithParams: GetObjectOutput;
     let stateAfterRemoveTask: GetObjectOutput;
@@ -36,10 +37,13 @@ describe('when calling org-formation perform tasks', () => {
         stateAfterDeploy2Targets = await s3client.getObject({Bucket: stateBucketName, Key: command.stateObject}).promise();
         spawnProcessMock = jest.spyOn(ChildProcessUtility, 'SpawnProcess');
 
-
         spawnProcessMock.mockReset();
         await PerformTasksCommand.Perform({...command, tasksFile: basePathForScenario + '1-deploy-serverless-workload-2targets.yml' });
         spawnProcessAfterRerunFileWithoutChanges = spawnProcessMock.mock;
+
+        spawnProcessMock.mockReset();
+        await PerformTasksCommand.Perform({...command, forceDeploy: true, tasksFile: basePathForScenario + '1-deploy-serverless-workload-2targets.yml' });
+        spawnProcessAfterRerunFileWithForceDeploy = spawnProcessMock.mock;
 
         spawnProcessMock.mockReset();
         await PerformTasksCommand.Perform({...command, tasksFile: basePathForScenario + '2-update-serverless-workload-with-parameters.yml' })
@@ -110,6 +114,12 @@ describe('when calling org-formation perform tasks', () => {
 
     test('after rerunning same serverless com task without changing, nothing got executed', () => {
         expect(spawnProcessAfterRerunFileWithoutChanges.calls.length).toEqual(0);
+    });
+
+    test('after rerunning same serverless com task without changing and force deploy, both targets get deployed', () => {
+        expect(spawnProcessAfterRerunFileWithForceDeploy.calls.length).toEqual(2);
+        expect(spawnProcessAfterDeploy2Targets.calls[0][0]).toEqual(expect.stringContaining('--region eu-central-1'));
+        expect(spawnProcessAfterDeploy2Targets.calls[1][0]).toEqual(expect.stringContaining('--region eu-central-1'));
     });
 
     test('after deploy 1 targets sls remove was called', () => {